third_party/WebKit/LayoutTests/fast/dom/shadow/focus-navigation-with-tabstop.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
<p>This tests TAB focus navigation with tabStop property on elements</p>
<pre id="console"></pre>
<script>

function prepareDOMTree(parent) {
    parent.appendChild(
        createDOM('div', {'id': 'testform'},
            createDOM('input', {'id': 'input-before'}),
            createDOM('div', {'id': 'host-div'},
                createShadowRoot(
                    createDOM('input', {'id': 'inner-input'}))),
            createDOM('input', {'id': 'input-after'})));
    parent.offsetTop;
}

var host_div;

function test() {
    debug("Testing shadow host with possible combinations of tabindex and tabStop");

    debug("Normal tab order without tabindex");

    host_div = document.getElementById("host-div");
    shouldBe('host_div.tabStop', 'false');

    var expectedOrder = [
      'input-before',
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order without tabindex but tabStop=true");
    host_div.tabStop = true;

    expectedOrder = [
      'input-before',
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=0 on host");

    host_div.tabIndex = 0;
    shouldBeEqualToString('host_div.getAttribute("tabindex")', '0');
    shouldBe('host_div.tabStop', 'true');

    expectedOrder = [
      'input-before',
      'host-div',
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=0 but tabStop = false on host");
    host_div.tabStop = false;

    expectedOrder = [
      'input-before',
      // 'host-div', // should skip host when tabStop = false
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=-1 on host");

    host_div.tabIndex = -1;
    shouldBeEqualToString('host_div.getAttribute("tabindex")', '-1');
    shouldBe('host_div.tabStop', 'false');

    expectedOrder = [
      'input-before',
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=-1 but tabStop=true on host");
    host_div.tabStop = true;

    expectedOrder = [
      'input-before',
      'host-div/inner-input',
      'input-after'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=1 on host");

    host_div.tabIndex = 1;
    shouldBeEqualToString('host_div.getAttribute("tabindex")', '1');
    shouldBe('host_div.tabStop', 'true');

    expectedOrder = [
      'input-before',
      'input-after',
      'host-div',
      'host-div/inner-input'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);

    debug("Normal tab order with tabindex=1 but tabStop=false on host");
    host_div.tabStop = false;

    expectedOrder = [
      'input-before',
      'input-after',
      // 'host-div', // should skip host when tabStop = false
      'host-div/inner-input'
    ];

    testFocusNavigationForward(expectedOrder);
    expectedOrder.reverse();
    testFocusNavigationBackward(expectedOrder);
}

function run_tests() {
    if (!window.eventSender) {
        testFailed('');
        return;
    }

    prepareDOMTree(document.body);
    test();

    debug('Test finished.');
}

run_tests();
</script>
</body>
</html>
